Documentation: A lot of new documentation has been added with LaTex, and a lot of old documentation has been removed from microsoft word. Much work is still needed with this.
Code:
The myDel.f90 dependency has been removed, and myDel.f90 has been deleted,
so now del.f90 is used.

gridGenTools.f90 was implemented and their corresponding routines have
been removed from griddata.f90.]

diffType is now computed internally in del.f90, and no longer exists externally.
gridType is now computed internally in interp.f90, and no longer exists externally.

The OO matlab scripts have been developed for the 1D relaxation as a proof of concept
for the 3D ADI.

The 3D ADI is still broken for Neumann BCs.

Next:
The ghost node is about to be introduced which will require changing.
	stencils.f90
	del.f90
	interp.f90
	applyBCs.f90
	coordinates.f90
Test cases will be run to make sure nothing is broken after and that
results are consistent with earlier versions. Cases to test include
	LDC, duct flow (for neumann BCs)

Finally, the MG will be developed for CC data and the projection cleaning
procedure will be investigated,implemented and tested to ensure this can
be used alongside the 3D ADI for insulating cases.