The poisson.f90 file was added to the unit testing folder. It solves the poisson equation for the following cases: 	Dirichlet 		Node data 			SOR 			ADI 			MG 		CC data 			SOR 			ADI 	Neumann 		Node data 			SOR 			ADI 			MG 		CC data 			SOR 			ADI
The only parameters that need adjusting are modelProblem in modelProblem_mod, which controls
the source term, boundary conditions, the exact solution etc., and s in the main program,
which controls whether the data is cell corner or cell center. If these two parameters disagree
an error is recieved.

This is a milestone, since the residuals of all tested problems go to zero within machine accuracy.
It should be clear from the setup in this file how source terms should be constructed.

There is still a little ambiguity about whether the BCs should be applied before or after taking the
mean of the source term. Although this may affect convergence rates, residuals were observed to reach
zero within machine accuracy for more than one setup.