Just successfully ran test_complexG.f90 using JAC_ops.f90 to solve a SS Poisson problem in 2 domains. The result between the 2 and 1 domain problems were identical, including the residual (Linf = 0.18351898123085392) of u_JAC vs exact. This is a great indication that the stitching was done properly.
One thing was done manually, and that was the interface conditions in applyBCs.f90.
This needs to be done in a systematic way, along with mesh.f90, and the stitch object
needs to be further developed to make this autonomous because grid IDs are needed, the
type of data that is being operated on along the direction it's being operated on, as well
as what's currently there (logicals indicating stitching is occuring).

Consider revising ops_split and stencils for these routines to be more similar to ops_del.
Fortunately, ops_split is working correctly because the staggered derivatives reach to ghost points.

Summary:
	Almost done stitching, need to implement BCs for interfaces along with mesh/grid
	Next
		Make sure JAC_ops is working for time stepping again
		Try to write JAC_ops for curl(curl) with variable material properties
		Run some simple test cases with this once the BC sim is done
		Work on higher level code.