This version is currently broken, but it has some great new features.
Just finished second BC sim, but cannot run momentum solver for for non-uniform grid.

It seems that uniform grids work, but everything defined in the mesh seems to indicate that
the mesh and stencils were defined correctly. BCs seem to remain satisfied, the solution blows
up very quickly. It must be something grid related since uniform grids work, on the other hand,
it must be a source term because BCs and ICs seem to stay satisfied.

About to switch back to a previous version to run the next BC sim, and maybe save some data that will be useful to debug the current version, e.g. L,D,U stencils for lapCC, etc. or maxval of some source terms.