InductionSolver is being given a makeover. Right now, induction_solver.f90 is being used and B & B0 are stored at cell face. The old computeJCrossB is being used with B,B0 and J that are interpolated to the CC. This seems to produce results consistent with previous simulations for BMC_102.
Next, the computeJCrossB must be exchanged for the correct jxB computed from raw interpolation
of J to the cell face and crossed with B (the new routine).

Once this is done, the old and unused variables should be removed and BMC_102 should be tested
one last time to make sure that consistent results remain.