Introduced name (character of length 1 ) to PCG for easier read during output.
Changed PCG_solver normalization from b to r0.

Fixed small bug with Nremaining in stop_clock.f90

Renamed N_mom to nstep_mom (as well as ind and nrg).

Currently, a jump occurs in the 2D LDC when trying to resolve the PPE and using N_mom=5.

It may be due to dividing r by r0 to escape (r and r0 are computed from div(u)).

Need to safely resolve this, then return to BC sims and play with parameters.