Just fixed apply_stitches_faces so that it does not reach to wall-coincident data, since this is defined by BCs. There are still likely scenarios that are not covered.
For example, stitching on wall-normal values is necessary for Neumann BCs (outflow), which is now not done from the fix above. The problem with allowing this is that Dirichlet BCs will not be satisfied on edges if apply_stitches_faces extends to the edge for wall-normal data.

This means that BC information is necessary for stitching for general cases.

In addition, some routines are not yet built to support wall-coincidient stitching, like sum(), which will double count on wall coincident data unless stitching information is passed in.