Fixed / generalized the apply_stitches_faces for the case of wall coincident data adjacent to a stitched face, which depends on stitching to adjacent faces and, in some cases, boundary conditions.
Running the flow over 2D square now with free-stream BCs on sides. Optimistic that this will work.

Implemented a "stop if NaNs" in the transient probe to avoid wasting resources.

Added a module with "fun" geometries including a full loop and figure 8 geometries, but all of them have yet to be tested/developed.

Next: compute energy budget for induction equation, brainstorm on results section of BC paper, and start additional BC cases.