Adjusted bounds in embed_extract (doesn't affect previous results).
Modified stop_clock and clock so that the last elapsed time is used
for estimates when elapsed time is returned negative.

Problem has been setup for lower Rem case, aside from that, still ready to run.
Try testing clock with negative value to make sure no errors exist when elapsed
time is negative (which occurs very late in the sim).