REMOVED THE STRANGEST COMPILATION ERROR I'VE EVER SEEN.
".exe" must appear after "TARGET = $(TARGET_DIR)$(PS)MOONS" in the makefile.
Otherwise, the the compiler will complain that object files
(with misspelled directories) do not exist and compilation will fail.

Also, moved all uncompiled files to a single folder (this did not
help the compilation, but is more clear to a user what is being used).