Just introduced 	iter_solver_params 	time_marching_params
to generalized
	nstep
	dt
	t
	N_max
for time marching and
	iter_max
	tol_relative
	tol_absolute
for iterative solvers.

This has greatly simplified the interface and will allow for simpler import / export.

It's now easier to adjust, on the fly, iterative solver params, such as relative tolerance, absolute tolerance, maximum number of iterations, and number of time steps for time marching solvers.

All of this was basically put into initialization, inside inputFile.f90

This was good.