2D hydro duct flow with Neumann outlet seems to work correctly.
The documentation needs to be updated because I think implicit Neumann BCs for PCG were implimented incorrectly. This does not affect any prior benchmarks since Neumann BCs were always performed explicitly.

It seems that the matrix A may not need to be explicitly symmetric, as is currently implemented, when multiplying a row of the equation satisfies symmetry.