Record breaking runtime for BMC_102 (6.259 minutes).
Just implemented:
	array
	sparse
	stencil

array is 1D and very simple / general.
sparse holds L,D,U for 1D
stencil holds L,D,U for 3D and will combine the diagonal for 3D stencils.

Need to think which stencils to store, and how they need to be constructed for scalar and vector fields.

Also implemented GF_laplacian GF_curl_curl and GF_interp, which assume the stencils are saved. They both look good, and general, but work needs to be done to send in the necessary grid_fields. First try the laplacian stencil, since this is simpler and can be more easily tested.

To store the stencils, consider saving them in block, since it has access to grid_field.