Running BMC_102 for finite Rem.
Improvements:
	Refinement occurs at a energy change tolerance: |d(data)/dt|/max(|data|) > tol.

refinement routines and code portions were moved to dynamic_refine_mesh.f90

Also to note, in case this wasn't stated before:
	boundary_flux is now working, and was tested against div(B).

To monitor ME in vacuum may require modification of this however, since
nhat is used and the integral of ME through the boundary is zero.

Need to use a time-dependent definition to compute ME flux through the surface over time.